Numeric.LongDouble

Description

This module re-exports the default platform-specfic ABI for C's long double.

Documentation

data LongDouble

The long double type on x86_64: 80bits of x87 FPU data in 128bits of space.

truncate' :: LongDouble -> LongDouble

Alternate versions of RealFrac methods that keep the value as a long double.

round' :: LongDouble -> LongDouble

Alternate versions of RealFrac methods that keep the value as a long double.

ceiling' :: LongDouble -> LongDouble

Alternate versions of RealFrac methods that keep the value as a long double.

floor' :: LongDouble -> LongDouble

Alternate versions of RealFrac methods that keep the value as a long double.

fromInt :: Int -> LongDouble

toInt :: LongDouble -> Int

fromDouble :: Double -> LongDouble

toDouble :: LongDouble -> Double